Refactor FileUtils class and add StringUtils class
The FileUtils class is refactored in this commit to improve its functionality. The `strip_non_ascii` method is renamed to `strip_non_ascii_filenames` for clarity. The method now renames files in the provided directory by removing non-ASCII characters. This helps to ensure that the renamed files don't overwrite existing files. The `StringUtils` class is also added, which provides a method `get_ascii_chars()` that returns a set of ASCII characters. This method is used in the refactored `strip_non_ascii_filenames` method to determine which characters should be considered ASCII. The commit also includes a version update in the setup.py file, from 0.1.1 to 0.1.2.