[weneed.utils.string_utils]: Refactor function `print_graphic` to `graphic_as_string` in StringUtils class
Refactor the function `print_graphic` in the `StringUtils` class to `graphic_as_string`. This function now accepts a list of texts instead of a single text. The function generates a graphic string by adding borders and formatting the texts. The width of the graphic string is controlled by the `width` parameter. The generated graphic string is returned as the result of this function.